从program中抽离接口调用代码
